# kestra/openapi.yml
openapi: 3.0.1
info:
  title: Kestra
  license:
    name: Apache 2.0
    url: https://raw.githubusercontent.com/kestra-io/kestra/master/LICENSE
  version: 1.0.0
...
paths:
...
  /api/v1/{tenant}/namespaces/{namespace}/inherited-secrets:
    get:
      tags:
      - Namespaces
      summary: List inherited secrets
      operationId: getInheritedSecrets
      parameters:
      - name: namespace
        in: path
        description: The namespace id
        required: true
        schema:
          type: string
      - name: tenant
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: getInheritedSecrets 200 response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: string
...
  /api/v1/{tenant}/namespaces/{namespace}/secrets:
    get:
      tags:
      - Namespaces
      summary: Get secrets for a namespace
      operationId: listNamespaceSecrets
      parameters:
      - name: namespace
        in: path
        description: The namespace id
        required: true
        schema:
          type: string
      - name: page
        in: query
        description: The current page
        required: true
        explode: false
        schema:
          type: integer
          format: int32
          default: 1
      - name: size
        in: query
        description: The current page size
        required: true
        explode: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: sort
        in: query
        description: The sort of current page
        explode: false
        schema:
          type: array
          nullable: true
          items:
            type: string
      - name: filters
        in: query
        description: Filters
        required: true
        explode: false
        schema:
          type: array
          items:
            $ref: "#/components/schemas/QueryFilter"
      - name: tenant
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: listNamespaceSecrets 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSecretListResponse_ApiSecretMeta_"
      deprecated: true
    post:
      tags:
      - Namespaces
      summary: Create or replace a namespace secret
      operationId: createNamespaceSecret
      parameters:
      - name: namespace
        in: path
        description: The namespace id
        required: true
        schema:
          type: string
      - name: tenant
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrUpdateNamespaceSecretRequest"
        required: true
      responses:
        "200":
          description: createNamespaceSecret 200 response
          content:
            application/json:
              schema:
                type: object
  /api/v1/{tenant}/namespaces/{namespace}/secrets/{key}:
    delete:
      tags:
      - Namespaces
      summary: Delete a namespace secret
      operationId: deleteNamespaceSecret
      parameters:
      - name: namespace
        in: path
        description: The namespace id
        required: true
        schema:
          type: string
      - name: key
        in: path
        description: The secret key
        required: true
        schema:
          type: string
      - name: tenant
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: deleteNamespaceSecret 200 response
          content:
            application/json:
              schema:
                type: object
    patch:
      tags:
      - Namespaces
      summary: Update a namespace secret
      operationId: updateNamespaceSecret
      parameters:
      - name: namespace
        in: path
        description: The namespace id
        required: true
        schema:
          type: string
      - name: key
        in: path
        description: The secret key
        required: true
        schema:
          type: string
      - name: tenant
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrUpdateNamespaceSecretRequest"
        required: true
      responses:
        "200":
          description: updateNamespaceSecret 200 response
          content:
            application/json:
              schema:
                type: object
  /api/v1/{tenant}/namespaces/{namespace}/settings:
    get:
      tags:
      - Namespaces
      summary: Get namespace settings
      operationId: getSettings
      parameters:
      - name: namespace
        in: path
        description: The namespace id
        required: true
        schema:
          type: string
      - name: tenant
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: getSettings 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NamespaceSettings"
    patch:
      tags:
      - Namespaces
      summary: Update namespace settings
      operationId: updateSettings
      parameters:
      - name: namespace
        in: path
        description: The namespace id
        required: true
        schema:
          type: string
      - name: tenant
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNamespaceSettingsRequest"
        required: true
      responses:
        "200":
          description: updateSettings 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NamespaceSettings"
...
  /api/v1/{tenant}/secrets:
    get:
      tags:
      - Secrets
      summary: Search secrets of all namespaces
      operationId: listSecrets
      parameters:
      - name: page
        in: query
        description: The current page
        required: true
        explode: false
        schema:
          type: integer
          format: int32
          default: 1
      - name: size
        in: query
        description: The current page size
        required: true
        explode: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: sort
        in: query
        description: The sort of current page
        explode: false
        schema:
          type: array
          nullable: true
          items:
            type: string
      - name: filters
        in: query
        description: Filters
        required: true
        explode: false
        schema:
          type: array
          items:
            $ref: "#/components/schemas/QueryFilter"
      - name: tenant
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: listSecrets 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSecretListResponse_ApiSecretMeta_"
...
components:
  schemas:
...
    ApiSecretListResponse_ApiSecretMeta_:
      required:
      - readOnly
      - results
      type: object
      properties:
        readOnly:
          type: boolean
        results:
          type: array
          items:
            $ref: "#/components/schemas/ApiSecretMeta"
        total:
          type: integer
          format: int64
    ApiSecretMeta:
      required:
      - key
      type: object
      properties:
        key:
          type: string
        description:
          type: string
        tags:
          type: object
          additionalProperties:
            type: string
...
    CreateOrUpdateNamespaceSecretRequest:
      required:
      - key
      type: object
      properties:
        key:
          minLength: 1
          type: string
          description: Secret key
        value:
          type: string
          description: Secret value (plain text; will be encrypted at rest)
          nullable: true
        description:
          type: string
          description: Optional description
          nullable: true
        tags:
          type: object
          additionalProperties:
            type: string
          description: "Optional tags (key-value map, or array of {key, value})"
          nullable: true
      description: Request body for creating or updating a namespace secret
...
    NamespaceSettings:
      required:
      - namespace
      - tenantId
      type: object
      properties:
        tenantId:
          type: string
        namespace:
          type: string
        description:
          type: string
        allowedNamespaces:
          type: array
          items:
            type: string
...
    QueryFilter:
      type: object
      properties:
        field:
          $ref: "#/components/schemas/QueryFilter.Field"
        operation:
          $ref: "#/components/schemas/QueryFilter.Op"
        value:
          type: object
    QueryFilter.Field:
      type: string
      enum:
      - QUERY
      - SCOPE
      - NAMESPACE
      - KIND
      - LABELS
      - METADATA
      - FLOW_ID
      - FLOW_REVISION
      - ID
      - ASSET_ID
      - TYPE
      - CREATED
      - UPDATED
      - START_DATE
      - END_DATE
      - EXPIRATION_DATE
      - STATE
      - TIME_RANGE
      - TRIGGER_EXECUTION_ID
      - TRIGGER_ID
      - TRIGGER_STATE
      - EXECUTION_ID
      - TASK_ID
      - TASK_RUN_ID
      - CHILD_FILTER
      - WORKER_ID
      - EXISTING_ONLY
      - MIN_LEVEL
      - PATH
      - PARENT_PATH
      - VERSION
      x-type: String
    QueryFilter.Op:
      type: string
      enum:
      - EQUALS
      - NOT_EQUALS
      - GREATER_THAN
      - LESS_THAN
      - GREATER_THAN_OR_EQUAL_TO
      - LESS_THAN_OR_EQUAL_TO
      - IN
      - NOT_IN
      - STARTS_WITH
      - ENDS_WITH
      - CONTAINS
      - REGEX
      - PREFIX
...
    UpdateNamespaceSettingsRequest:
      type: object
      properties:
        description:
          type: string
          description: Namespace description
          nullable: true
        allowedNamespaces:
          type: array
          description: List of namespace ids allowed to access this namespace; null
            or empty means all allowed
          nullable: true
          items:
            type: string
      description: Request body for updating namespace settings
...